EditIndexCard, FAB in FirstFragment, nav_graph
Created a new Fragement for the creation of an index card.
Then used the navigation component to add a navigation action from the first fragment to the new editIndexCards one.
I had to move the FAB to the First Fragment because a navigation from the Main Act to a fragment was not possible.