new Custom Dlg from DialogFragment
Created a new custom dlg  that extends from the DialogFragment class and implements an OnClickListener for the two custom buttons. I also created a Background layout for the custom dlg and set it to it with the dlg.getWindow().setBackgroundDrawable(getDrawable(getContext(), R.drawable.dlg_background)); in the overriden onCreateView method.
Insted of using the onCreateDialog() method like in the alertDialog dlg I overrid the OnCreateView (to inflate the layout for the dlg) and the onViewCreated() (to get views like edittexts buttons etc. and to set the title or the background for the dilalog). This custom dialog is created more like a fragment/activity insted of using the builder class for the AppCompatDialogFragment dlg.

I as well changed the layout of the category_spinner_adapter , fragment edit index card and used the viewmodel class inside the dlg to write a new category to the room db.