Right on Click and navigate from parent to child
In EditIndexCardFrg I added the child fragment in to a frame layout by using transaction and the getChildFragmentManager().beginTransaction();
In EditIndexCardFrgChild I now Properly use the onClickListener for my swicthAQ button. and call

 EditIndexCardFrgDirections.ActionEditIndexCardToEditIndexCardFrgChild action =
                        EditIndexCardFrgDirections.actionEditIndexCardToEditIndexCardFrgChild("Test");

                NavHostFragment.findNavController(EditIndexCardFrgChild.this)
                        .navigate(action);

I call it from the parent fragment because the dastination is till the parent fragment even if I call this from the child fragment... I can then go to the child fragment but it opens as a new stand alone fragment and not inside the parent as intende.

Next Step is to figure out if it is posible to use the nav components to navigate only between the children (maybe with a nested graph ) or if it will be simpler to use the transaction and change the child fragents over the parent itself and migrate the button from the child to the parent. (add animation to transaction and swith fragments in the frame layout)