IndexCardViewModel / getPrimaryNavigationFrg
Created a IndexCardViewModel with only one function to write index card into the room db
In the MainActivty I needed the "active" EditIndexCardFrg to have acces to the safeIndexCard() method in the new fab.onClick method specialy for this Fragment.
With the hostFragment.getChildFragmentManager().getPrimaryNavigationFragment() I can acces the right instance because the EditIndexCardFrg in this case is the last fragment I navigated to.